+++ Js2IL.Tests/ArrowFunction/JavaScript/ArrowFunction_LexicalThis_ConstructorAssigned.js
@@ -1,0 +1,12 @@
+class Counter {
+    constructor(initial) {
+        this.x = initial;
+        this.getX = () => this.x;
+    }
+}
+
+const c = new Counter(5);
+const other = { x: 9, getX: c.getX };
+
+console.log(c.getX());
+console.log(other.getX());